+++ cmd_plot.py
@@ -1,4 +1,6 @@
 from pathlib import Path
+import shutil
+from feasytools import ArgChecker
 from v2sim import *
 
 def clear_all(p:str):
