+++ gui_para.py
@@ -1,9 +1,14 @@
 import tkinter as tk
 import tkinter.ttk as ttk
 import multiprocessing as mp
+import time
+import sys
+from pathlib import Path
+from typing import Optional
 from tkinter import filedialog as fd
 from tkinter import messagebox as mb
 from fgui import ScrollableTreeView, LogItemPad, EditMode
+from feasytools import time2str
 from v2sim import *
 
 _L = CustomLocaleLib.LoadFromFolder("./resources/gui_para")
