+++ requirements.txt
@@ -1,7 +1,7 @@
 libsumo
 numpy
 matplotlib
-feasytools>=0.0.28
+feasytools>=0.0.30
 fpowerkit>=0.3.2
 pyproj
 requests
