+++ v2sim/__init__.py
@@ -4,4 +4,10 @@ from .trafficgen import *
 from .plotkit import *
 from .plugins import *
 from .statistics import *
-from .sim_core import *
+from .sim_core import (
+    load_external_components,
+    get_sim_params,
+    simulate_multi,
+    simulate_single,
+    V2SimInstance,
+)
