+++ v2sim/plotkit/__init__.py
@@ -1,2 +1,2 @@
-from .plot import *
+from .plot import AdvancedPlot
 from .reader import StatisticsNotSupportedError, ReadOnlyStatistics
