+++ v2sim/traffic/cslist.py
@@ -1,8 +1,8 @@
 from itertools import repeat
-from typing import Iterable, Type, Union, Optional, TypeVar, Generic
+from typing import Iterable, Type, Union, Optional, TypeVar, Generic, List
 from feasytools import RangeList, KDTree, Point
 from ..locale import Lang
-from .utils import readXML
+from .utils import ReadXML
 from .params import *
 from .evdict import EVDict
 from .cs import CS, SCS, FCS
@@ -11,10 +11,10 @@ T_CS = TypeVar("T_CS", FCS, SCS)
 CS_Type = Type[T_CS]
 
 
-def _LoadCSList(filePath:str, csType:CS_Type) -> list:
+def _LoadCSList(filePath:str, csType:CS_Type) -> List[CS]:
     assert csType == FCS or csType == SCS
-    _cs:list[CS] = []
-    root = readXML(filePath).getroot()
+    _cs:List[CS] = []
+    root = ReadXML(filePath).getroot()
     if root is None: raise ValueError("Invalid CS file")
     for cs_node in root:
         par_pbuy = ()
